fix: 데이터 타입 변경

+++ lib/features/workout_history/presentation/screens/workout_statics_screen.dart
@@ -588,7 +588,7 @@ class _WorkoutStaticsScreenState extends ConsumerState<WorkoutStaticsScreen> {
   }
 
   Widget _buildElevationItem() {
-    final int? elevation = _currentStatistics?.summary.totalElevationGain;
+    final double? elevation = _currentStatistics?.summary.totalElevationGain;
     return Expanded(
       child: InfoItem(
         label: '상승 고도',
